Add helper functions to retrieve SHA256 checksums
Checksums are important for easily detecting whether a given file has
changed or not, and whether the file has been corrupted. Store this
information as part of the bindata payload, so we can return it
without asking users to 1) make a copy of the file in memory 2)
checksum it.

In the future, I'd like to checksum each file once before returning
it, though this might be tricky to implement.

Add two new API's for retrieving files as strings.

Fixes #1.