convert: try to make naming clearer
We refer to a value as a "dir" when it can be either a directory or
a plain file, which is confusing and led to an error.

There's another problem where a value can either be an absolute path
or a relative path, which is concerning. I tried to make it absolute
in all cases but this led to errors in the generated filename. Will
try to fix later.

Add several tests which document the incorrect behavior that's fixed
by #22.