release: slow path in more cases
If the string contains a character in the Unicode format range,
staticcheck recommends escaping it instead of embedding it as is.
Detect characters in this range and trigger the slow path if we hit
one. Should not be too much slower than the existing code since we are
already doing a pass over the range previously.