all: replace "Debug" with "AssetDebug"
Josh Bleecher Snyder made the point that if this file is compiled
in with other files in a package, "Debug" is a very generic word.
"AssetDebug" is less likely to have conflicts with existing packages.

This is a breaking change, which I'm not happy about, but it's a new
one, so hopefully very few people are relying on the behavior that was
recently added. Add a "CHANGELOG" to helpfully make this change more
clear to people.